Organized files, created NewspaperContext to link to database. Created Migration.
