+++ Newspaper/Models/Article.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Newspaper
+namespace Newspaper.Models
 {
     public class Article
     {
